requirements.txt
accelerate==1.8.1
dotenv==0.9.9
easyocr==1.7.1
faiss==1.7.2
langchain==0.3.26
langchain-anthropic==0.3.16
langchain-community==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.26
langchain-text-splitters==0.3.8
langgraph==0.5.0
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.1
langgraph-sdk==0.1.71
langsmith==0.4.1
Markdown==3.8.2
matplotlib==3.8.4
mss==9.0.1
nltk==3.9.1
numpy==1.26.4
openai==1.90.0
opencv-python-headless==4.11.0.86
pandas==2.3.0
pillow
psutil==5.9.8
pyautogui==0.9.54
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller
pynput==1.7.6
PySide6==6.7.0
pytesseract==0.3.10
python-dotenv==1.1.0
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==2.7.0
tiktoken==0.9.0
torch==2.2.2
torchao==0.11.0
torchvision==0.17.2
tornado
tqdm
transformers==4.41.2
watchdog==6.0.0
yarl==1.20.1
